+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-24 10:39:52
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-06-25 11:46:42
+// Commit: feat(test): add input validation and sanitization for test
 
 class testComponentState {
-  final String commitRef = '2248';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-06-24T10:39:52+05:00');
+  final String commitRef = '2293';
+  final String lastMessage = 'feat(test): add input validation and sanitization for test';
+  final DateTime timestamp = DateTime.parse('2026-06-25T11:46:42+05:00');
 }